Copy over Perceiver encoding
This will be changed more, but serves as a baseline. I don't particularly want to add perceiver-model as a dependency for this repo, so having some code duplication is probably more acceptable. Additionally, since this will be extended and modified more than the version in perceiver-model, I think it should be okay